fix: updated entities to match the db schema, for order submission

+++ trading-app/src/main/java/com/leap/leaplaughlove/trading/order/OrderHistoryItem.java
@@ -1,6 +1,5 @@
 package com.leap.leaplaughlove.trading.order;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -19,7 +18,7 @@ public record OrderHistoryItem(
         UUID orderId,
         String symbol,
         String side,
-        BigDecimal quantity,
+        Long quantity,
         String status,
         OffsetDateTime submittedAt,
         OffsetDateTime filledAt
